test(*): get the initial test suite passing
There were two reasons the tests were failing initially; (1) the dash shell (`/bin/sh`) does not
have the "pipefail" option, and (2) the child process was leaking file descriptors. These are both
resolved now by (1) removing the pipefail option and (2) ensuring that if the out/err pipes are
present on the child process that they are closed and that the process as a whole is also closed.