fix(exec, bundle): allow passing one exec result to another and bandaid fix the bundle command
Previously only numbers and strings could be passed as exec args, but now you can pass the results
of one exec call to another and in this case, only the sdtout will be used from the first result.
Also "fix" an intermittent failure of the bundle command which is caused by a non-deterministic key
order for the `result.files` array when source maps were emitted - by disabling source maps, we
effectively ensure that the first entry is always the file we want.