feat(cli): add the ability to run markdown files
The feature of https://github.com/google/zx to run markdown files as scripts is incredibly helpful
as it provides the ability to include nice looking and well structured documentation right alongside
your scripts. This was one of the largest outstanding features of zx which was absent from dzx, and
with this code in place, now dzx has the ability to run markdown files as scripts as well. See the
example markdown file for detailed documentation.